Import of DURIST/Proc-ProcessTable-0.35 from CPAN.
gitpan-cpan-distribution: Proc-ProcessTable
gitpan-cpan-version:      0.35
gitpan-cpan-path:         DURIST/Proc-ProcessTable-0.35.tar.gz
gitpan-cpan-author:       DURIST
gitpan-cpan-maturity:     released

+++ os/Linux.h
@@ -57,7 +57,7 @@ struct procstat {
   int priority;
   unsigned timeout;
   unsigned itrealvalue;
-  int starttime;
+  unsigned long  starttime;
   unsigned vsize;
   unsigned rss;
   unsigned rlim;
